Fix multi-line message flicker on resize
So by default NSTextView flickers if most of text is out of viewport bounds, so I had to remove my throttle on our existing logic to battle against this. But it comes with performance cost of too many layouts.

As a quick fix I switch to less performant method when message content height is below 50px.