Disable sorting when adding messages
This fixes a nasty bug I spent 1.5hrs debugging. When we sort the indexes can become stale thus we won't be updating correct message rows.

We need to find a way around this at some point as there is a possiblity we may want to insert messages not at the tail or head.