macos: fix placeholder bug with chinese input
User reported Chinese users still see the placeholder when they start typing in Chinese characters.
So apparently there is a feature in macOS for these languages called Chinese IME (Input Method Editor) which lays
out text temporarily without committing it. This method can detect this and hide the placeholder. And show it back
when that text is removed.